refactor: reorganize controllers, implement RESTful routing, and split large CSS files
