--- app/views/Landing/gallery.php
@@ -1,270 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Artvault - Gallery</title>
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Cinzel:wght@400;600;700&family=Lato:wght@300;400;700&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="/css/index.css">
-</head>
-<body>
-
-    <!-- NAVBAR -->
-    <nav class="navbar">
-        <div class="nav-logo">
-            <img src="/img/logo/Artvault.png" alt="Artvault Logo" class="logo-img">
-        </div>
-
-        <ul class="nav-menu">
-            <li><a href="/" class="">Home</a></li>
-            <li><a href="/gallery" class="active">Gallery</a></li>
-            <li><a href="/contact">Contact</a></li>
-            <li><a href="/about">About</a></li>
-        </ul>
-
-        <div class="nav-actions">
-            <button class="btn btn-login" onclick="location.href='/login'">Log In</button>
-            <button class="btn btn-signup" onclick="location.href='/signup'">Sign Up</button>
-        </div>
-    </nav>
-
-    <!-- GALLERY -->
-    <div class="gallery-page">
-
-        <div class="gallery-topbar">
-            <button class="btn-category">Category</button>
-            <div class="search-box">
-                <input type="text" placeholder="Search....">
-                <button class="search-btn">🔍</button>
-            </div>
-        </div>
-
-        <div class="gallery-grid">
-            <a href="/art/1" class="art-card">
-                <img src="/img/gallery/Claymonster.png" alt="Claymonster">
-                <div class="art-info">
-                    <p class="art-title">Claymonster</p>
-                    <p class="art-author">Made by : Chisa Evelyn</p>
-                    <p class="art-like"><img src="/img/icon/like.png" class="Art-like-img">25</p>
-                </div>
-            </a>
-            <a href="" class="art-card">
-                <img src="/img/gallery/Koi-Pond.png" alt="Koi Pond">
-                <div class="art-info">
-                    <p class="art-title">Koi Pond</p>
-                    <p class="art-author">Made by : Odin Madun</p>
-                    <p class="art-like"><img src="/img/icon/like.png" class="Art-like-img">49</p>
-                </div>
-            </a>
-            <a href="/landing/detail" class="art-card">
-                <img src="/img/gallery/Trash-Hunt.png" alt="Trash Hunt">
-                <div class="art-info">
-                    <p class="art-title">Trash Hunt</p>
-                    <p class="art-author">Made by : Viktor Wembu</p>
-                    <p class="art-like"><img src="/img/icon/like.png" class="Art-like-img">78</p>
-                </div>
-            </a>
-            <a href="" class="art-card">
-                <img src="/img/gallery/Cherish-the-moment.png" alt="Cherish the moment">
-                <div class="art-info">
-                    <p class="art-title">Cherish the moment</p>
-                    <p class="art-author">Made by : Daniel Caesar</p>
-                    <p class="art-like"><img src="/img/icon/like.png" class="Art-like-img">65</p>
-                </div>
-            </a>
-            <a href="" class="art-card">
-                <img src="/img/gallery/Ayo-punya-cita-cita.png" alt="Ayo punya cita-cita">
-                <div class="art-info">
-                    <p class="art-title">Ayo punya cita-cita</p>
-                    <p class="art-author">Made by : Faysal Pratama Agung</p>
-                    <p class="art-like"><img src="/img/icon/like.png" class="Art-like-img">69</p>
-                </div>
-            </a>
-            <a href="" class="art-card">
-                <img src="/img/gallery/Billie-Eilish.png" alt="Billie Eilish">
-                <div class="art-info">
-                    <p class="art-title">Billie Eilish</p>
-                    <p class="art-author">Made by : Depon Vintjai</p>
-                    <p class="art-like"><img src="/img/icon/like.png" class="Art-like-img">72</p>
-                </div>
-            </a>
-            <a href="" class="art-card">
-                <img src="/img/gallery/A-positive-spin-on.png">
-                <div class="art-info">
-                    <p class="art-title">A positive spin on n...</p>
-                    <p class="art-author">Made by : Tang You Hoong</p>
-                    <p class="art-like"><img src="/img/icon/like.png" class="Art-like-img">51</p>
-                </div>
-            </a>
-            <a href="" class="art-card">
-                <img src="/img/gallery/Menggapai-Indonesia.png" alt="Menggapai Indonese...">
-                <div class="art-info">
-                    <p class="art-title">Menggapai Indonese...</p>
-                    <p class="art-author">Made by : Marcello Adil</p>
-                    <p class="art-like"><img src="/img/icon/like.png" class="Art-like-img">91</p>
-                </div>
-            </a>
-            <a href="" class="art-card">
-                <img src="/img/gallery/Vibrant-River.png" alt="Vibrant River">
-                <div class="art-info">
-                    <p class="art-title">Vibrant River</p>
-                    <p class="art-author">Made by : Reyfan Andika</p>
-                    <p class="art-like"><img src="/img/icon/like.png" class="Art-like-img">34</p>
-                </div>
-            </a>
-            <a href="" class="art-card">
-                <img src="/img/gallery/Anonymous-Protagonis.png" alt="Anonymous Protagonis">
-                <div class="art-info">
-                    <p class="art-title">Anonymous Protag...</p>
-                    <p class="art-author">Made by : Felicia Chiao</p>
-                    <p class="art-like"><img src="/img/icon/like.png" class="Art-like-img">101</p>
-                </div>
-            </a>
-            <a href="" class="art-card">
-                <img src="/img/gallery/A-Chill-Doomsday.png" class="art-card" alt="A Chill Doomsday">
-                <div class="art-info">
-                    <p class="art-title">A Chill Doomsday</p>
-                    <p class="art-author">Made by : Nicholas Jo</p>
-                    <p class="art-like"><img src="/img/icon/like.png" class="Art-like-img">99</p>
-                </div>
-            </a>
-            <a href="" class="art-card">
-                <img src="/img/gallery/Melody-in-Guitar.png" alt="Melody in Guitar">
-                <div class="art-info">
-                    <p class="art-title">Melody in Guitar</p>
-                    <p class="art-author">Made by : Jo Halimawan</p>
-                    <p class="art-like"><img src="/img/icon/like.png" class="Art-like-img">120</p>
-                </div>
-            </a>
-        </div>
-
-         <div class="add-work-wrap">
-              <button class="btn-add-work" onclick="openAddWork()">+ Add your work!</button>
-                    <!-- POPUP ADD WORK -->
-        <div class="add-work-overlay" id="addWorkOverlay">
-            <div class="add-work-popup">
-
-                <div class="add-work-header">
-                    <img src="img/logo/artvault.png" alt="Artvault Logo" class="add-work-logo">
-                    <span>Come on, submit your interesting work to be exhibited!</span>
-                    <img src="/img/icon/user.png" alt="User" class="user-icon">
-                </div>
-
-                <div class="add-work-body">
-                <button class="btn-back-popup" onclick="closeAddWork()">
-                        <img src="/img/icon/back.png" alt="Back" class="back-icon-2">
-                    </button>
-                    <!-- Upload Area -->
-                    <div class="upload-area" onclick="document.getElementById('fileInput').click()">
-                        <input type="file" id="fileInput" accept="image/*" style="display:none" onchange="previewImage(event)">
-                        <img id="previewImg" src="" alt="" style="display:none; width:100%; height:100%; object-fit:cover; border-radius:8px;">
-                        <button class="btn-upload" id="uploadBtn">Upload your art</button>
-                    </div>
-
-                    <!-- Name -->
-                    <div class="add-work-field">
-                        <input type="text" placeholder="Add the name of your Art!">
-                    </div>
-
-                    <!-- Description -->
-                    <div class="add-work-field">
-                        <textarea placeholder="Add your description!" rows="3"></textarea>
-                    </div>
-
-                    <!-- Submit -->
-                    <button class="btn-add-submit">Add</button>
-
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-
-    <!-- FOOTER -->
-    <div class="box-2"></div>
-    <footer class="footer">
-        <div class="footer-brand">
-            <div class="footer-brand-top">
-                <img src="/img/logo/Artvault-white.png" alt="Artvault Logo" class="footer-logo">
-                <span class="footer-brand-name">Artvault</span>
-            </div>
-        </div>
-        <div class="footer-main">
-            <div class="footer-col">
-                <h4 class="footer-heading">Navigation</h4>
-                <ul class="footer-links">
-                    <li><a href="/">Home</a></li>
-                    <li><a href="/gallery">Gallery</a></li>
-                    <li><a href="/about">About</a></li>
-                    <li><a href="/contact">Contact</a></li>
-                </ul>
-            </div>
-            <div class="footer-col">
-                <h4 class="footer-heading">Contact</h4>
-                <ul class="footer-contact">
-                    <li>
-                        <span class="contact-icon"><img src="/img/icon/mail.png" alt="Email" class="contact-icon-img"></span>
-                        <span>Artvault@gmail.com</span>
-                    </li>
-                    <li>
-                        <span class="contact-icon"><img src="/img/icon/location.png" alt="Location" class="contact-icon-img"></span>
-                        <span>Pontianak, Kalimantan Barat, Indonesia</span>
-                    </li>
-                    <li>
-                        <span class="contact-icon"><img src="/img/icon/telephone.png" alt="Phone" class="contact-icon-img"></span>
-                        <span>+62 897 3871 170</span>
-                    </li>
-                </ul>
-            </div>
-            <div class="footer-col">
-                <h4 class="footer-heading">Help & Center</h4>
-                <ul class="footer-links">
-                    <li><a href="#">Customer Support</a></li>
-                    <li><a href="#">Terms and Services</a></li>
-                    <li><a href="#">Privacy Policy</a></li>
-                    <li><a href="#">Cancellation and Refund Policy</a></li>
-                </ul>
-            </div>
-            <div class="footer-col footer-social-col">
-                <img src="/img/logo/Artvault-white.png" alt="Artvault" class="footer-logo-big">
-                <div class="footer-social">
-                    <p class="footer-follow">Follow Us</p>
-                    <div class="social-icons">
-                        <a href="#"><img src="/img/icon/instagram.png" alt="Instagram"></a>
-                        <a href="#"><img src="/img/icon/facebook.png" alt="Facebook"></a>
-                        <a href="#"><img src="/img/icon/tiktok.png" alt="TikTok"></a>
-                        <a href="#"><img src="/img/icon/youtube.png" alt="YouTube"></a>
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div class="footer-bottom">
-            <p>Copyright &copy; 2025 Artvault All Rights Reserved</p>
-        </div>
-    </footer>
-    <script>
-            function openAddWork() {
-        document.getElementById('addWorkOverlay').classList.add('active');
-    }
-
-    function closeAddWork() {
-        document.getElementById('addWorkOverlay').classList.remove('active');
-    }
-
-    function previewImage(event) {
-        const file = event.target.files[0];
-        if (!file) return;
-        const preview = document.getElementById('previewImg');
-        const uploadBtn = document.getElementById('uploadBtn');
-        preview.src = URL.createObjectURL(file);
-        preview.style.display = 'block';
-        uploadBtn.style.display = 'none';
-        
-    }
-    </script>
-    
-    <script src="/js/script.js"></script>
-</body>
-</html>
