+++ app/views/author/gallery.php
